requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.9
django-adminlte-ui==1.5.1
django-autocomplete-light==3.9.4
django-ckeditor==6.3.0
django-crispy-forms==1.14.0
django-dal==1.0.0
django-embed-video==1.4.2
django-import-export==2.6.1
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-mptt==0.13.4
django-richtextfield==1.6.1
django-search-listview==0.2.1
django-select2-forms==3.0.0
django-social-share==2.3.0
django-sortedm2m==3.1.1
django-treebeard==4.3
django-widget-tweaks==1.4.9
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
psycopg2==2.9.3
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
PyYAML==6.0
reportlab==3.6.6
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
tzdata==2021.5
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.5
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
